More Code Cleanup
- Removed imgui folder in main since it was a duplicate of the one in libraries (?)
- Renamed testinput to ControllerInput
- Moved OBJ_Loader into boilerplate
- Fixed a multitude of type cast warnings

+++ 585Game/src/Boilerplate/Text.cpp
@@ -34,7 +34,7 @@ std::map<char, Character> initFont(const char* font) {
 		generateChar(face),
 		glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
 		glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-		face->glyph->advance.x
+		(unsigned int)face->glyph->advance.x
 		};
 		Characters.insert(std::pair<char, Character>(c, character));
 	}
